🎨 fix responsive
