+++ src/app/scoreboards/basketball/component/BasketballTeamScore.tsx
